requirements.txt
atlassian-python-api>=3.41.0
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0
schedule>=1.2.0
